+++ scene/lobby_scene.h
@@ -5,7 +5,7 @@
 class LobbyScene : public Scene {
 public:
     void Init() override;
-    void Update(float deltaTime) override;
+    void Update(double deltaTime) override;
     void Render() override;
     void Shutdown() override;
     static void StartGame();
