+++ scene/main_scene.h
@@ -4,7 +4,7 @@
 class MainScene : public Scene {
 public:
     void Init() override;
-    void Update(float deltaTime) override;
+    void Update(double deltaTime) override;
     void Render() override;
     void Shutdown() override;
 };
